Move types related to parameters from src/parameters/mod.ts to src/parameters/types.ts
