+++ src/functions/interactivity/block_actions_router_test.ts
@@ -15,7 +15,7 @@ import type {
 import type {
   ParameterSetDefinition,
   PossibleParameterKeys,
-} from "../../parameters/mod.ts";
+} from "../../parameters/types.ts";
 import type { SlackFunctionDefinition } from "../mod.ts";
 import { DefineFunction, Schema } from "../../mod.ts";
