+++ src/functions/tester/types.ts
@@ -1,7 +1,7 @@
 import type {
   ParameterSetDefinition,
   PossibleParameterKeys,
-} from "../../parameters/mod.ts";
+} from "../../parameters/types.ts";
 import type { SlackFunctionDefinition } from "../mod.ts";
 import type {
   FunctionContext,
