+++ tests/integration/parameters/parameter_variable_test.ts
@@ -1,5 +1,6 @@
 import { DefineProperty } from "../../../src/parameters/define_parameter.ts";
-import { ParameterVariable, SingleParameterVariable } from "../../../src/parameters/mod.ts";
+import { ParameterVariable } from "../../../src/parameters/mod.ts";
+import { SingleParameterVariable } from "../../../src/parameters/types.ts";
 import { DefineType } from "../../../src/types/mod.ts";
 import SchemaTypes from "../../../src/schema/schema_types.ts";
 import {
